CATROID-1087 "Delete look" and "Delete background" do not work with looks from the web

+++ catroid/src/main/java/org/catrobat/catroid/ui/recyclerview/controller/LookController.java
@@ -61,7 +61,9 @@ public class LookController {
 	}
 
 	public void delete(LookData lookToDelete) throws IOException {
-		StorageOperations.deleteFile(lookToDelete.getFile());
+		if (lookToDelete.getFile().exists()) {
+			StorageOperations.deleteFile(lookToDelete.getFile());
+		}
 	}
 
 	public LookData pack(LookData lookToPack) throws IOException {
